added name changes for conformance with skeleton team requierments, also added a bug fix in main activity, for when contact field in empty, and we try to find a contact within it (when adding or removing a favorite)

+++ project/src/com/technion/coolie/teletech/MainActivity.java
@@ -147,18 +147,30 @@ public class MainActivity extends CoolieActivity implements
 
 	void addToFavourites() {
 		// TODO CHANGE LATER
-		int position = FullContactInformation.position();
-		ContactInformation contact = MainActivity.contacts.get(position);
-		contact.setFavourite(true);
-		db.insertFavourite(contact);
+
+		final int position = FullContactInformation.position();
+
+		// added if for OutOfBoundsException
+		if (position < MainActivity.contacts.size()) {
+			final ContactInformation contact = MainActivity.contacts
+					.get(position);
+			contact.setFavourite(true);
+			db.insertFavourite(contact);
+		}
+
 	}
 
 	void removeFromFavourites() {
 		// TODO CHANGE LATER
-		int position = FullContactInformation.position();
-		ContactInformation contact = MainActivity.contacts.get(position);
-		contact.setFavourite(false);
-		db.deleteFavourite(contact.ID().toString());
+
+		final int position = FullContactInformation.position();
+		// added if for OutOfBoundsException
+		if (position < MainActivity.contacts.size()) {
+			final ContactInformation contact = MainActivity.contacts
+					.get(position);
+			contact.setFavourite(false);
+			db.deleteFavourite(contact.ID().toString());
+		}
 	}
 
 	@Override
